// frontend/web/src/services/API/proserve-wb-publishing-api/models/Product.ts
export interface Product {
    /**
     * The reference to the associated project.
     * @type {string}
     * @memberof Product
     */
    projectId: string;
    /**
     * Unique ID of the product.
     * @type {string}
     * @memberof Product
     */
    productId: string;
    /**
     * The reference to the associated technology id.
     * @type {string}
     * @memberof Product
     */
    technologyId: string;
    /**
     * The reference to the associated technology name.
     * @type {string}
     * @memberof Product
     */
    technologyName: string;
    /**
     * Status of the product (CREATING|CREATED|FAILED|PAUSED|ARCHIVING|ARCHIVED).
     * @type {string}
     * @memberof Product
     */
    status: string;
    /**
     * Name of the product.
     * @type {string}
     * @memberof Product
     */
    productName: string;
    /**
     * Type of the product (WORKBENCH).
     * @type {string}
     * @memberof Product
     */
    productType: string;
    /**
     * Description of the product.
     * @type {string}
     * @memberof Product
     */
    productDescription?: string;
    /**
     * Recommended version ID of the product.
     * @type {string}
     * @memberof Product
     */
    recommendedVersionId?: string;
    /**
     * 
     * @type {Array<string>}
     * @memberof Product
     */
    availableStages?: Array<string>;
    /**
     * 
     * @type {Array<string>}
     * @memberof Product
     */
    availableRegions?: Array<string>;
    /**
     * Creation date of the product.
     * @type {string}
     * @memberof Product
     */
    createDate: string;
    /**
     * Last update date of the product.
     * @type {string}
     * @memberof Product
     */
    lastUpdateDate: string;
    /**
     * User that created the product.
     * @type {string}
     * @memberof Product
     */
    createdBy: string;
    /**
     * User that performed the last update on the product.
     * @type {string}
     * @memberof Product
     */
    lastUpdatedBy: string;
    /**
     * 
     * @type {Array<VersionSummary>}
     * @memberof Product
     */
    versions?: Array<VersionSummary>;
}