// frontend/web/src/services/API/proserve-wb-publishing-api/models/VersionSummary.ts
export interface VersionSummary {
    /**
     * The id of the version
     * @type {string}
     * @memberof VersionSummary
     */
    versionId: string;
    /**
     * The version number
     * @type {string}
     * @memberof VersionSummary
     */
    name: string;
    /**
     * The description for the version
     * @type {string}
     * @memberof VersionSummary
     */
    description?: string;
    /**
     * The version type (RELEASE_CANDIDATE|RELEASED|RESTORED)
     * @type {string}
     * @memberof VersionSummary
     */
    versionType: string;
    /**
     * 
     * @type {Array<string>}
     * @memberof VersionSummary
     */
    stages: Array<string>;
    /**
     * Status of the version summary (CREATED|FAILED|RETIRED|PROCESSING).
     * @type {string}
     * @memberof VersionSummary
     */
    status: string;
    /**
     * Flag that indicates if version is recommended
     * @type {boolean}
     * @memberof VersionSummary
     */
    recommendedVersion: boolean;
    /**
     * Timestamp of when the last change happened to the version
     * @type {string}
     * @memberof VersionSummary
     */
    lastUpdate: string;
    /**
     * User ID of the person who last changed the version
     * @type {string}
     * @memberof VersionSummary
     */
    lastUpdatedBy?: string;
    /**
     * Name of the version that this version is restored from, if it is restored
     * @type {string}
     * @memberof VersionSummary
     */
    restoredFromVersionName?: string;
    /**
     * AMI ID of the original image
     * @type {string}
     * @memberof VersionSummary
     */
    originalAmiId?: string;
}